remove __main__
